initial work on a new version of the tray that takes up the whole box
The player coin counters were not working well because of size
restrictions and deficient weight.  Besides, players being able to hide
their coins may be a fun part of the game for some people.

The new tray design will work as a play area.  So players merely remove
the lid, shuffle the cards, replace them in their original location, and
play the game.