increase space for the coin tray
The sides of the coin tray are not symmetric. The card side is
perpendicular to the bottom.  And the far side has a more pronounced
angle.  The net change in volume is positive.

The card ramp has been made slightly steeper. This gives the coin tray
a little more volume.

Finally, the coin tray was made slightly narrower.  As rather intense
sanding is still required to fit the print into the game box.
Sanding may still be required.